Move config out of src (#683)
Why:
Currently our build process includes certain files that are expected to
be overwritten with files from a remote repo. Some of these files are in
our src/ folder which makes git management a pain when testing various
builds locally and it is also not intuitive to new contributors during
their on boarding how these files are intended to be used. We would like
to make this process more clear and easier to maintain.

This commit:
Moves the configurable files out of ./src/configuration/ and into
./config as to better separate from the source code.
