+++ bin/download_brand_colors_methods.rb
@@ -2,7 +2,7 @@
 require_relative "./helpers"
 
 BRAND_COLORS_FILE_NAME = "brandColors.ts"
-BRAND_COLORS_FILE_PATH = "src/configuration"
+BRAND_COLORS_FILE_PATH = "config"
 
 def download_brand_colors_file(ha_label, access_token)
   brand_colors_file_url = "https://#{access_token}@raw.githubusercontent.com/Path-Check/pathcheck-mobile-resources/#{mobile_resources_commit}/brand_colors/#{HA_LABEL}/#{BRAND_COLORS_FILE_NAME}"
